Correções após rebase
